# Implemented nested write transactions for realm.
Problem:

The current implementation of the method [write:block] in the MJDRealmCacheService.m class doesn't support nested transactions because after the block() is executed, the code commit the transaction directly.

Solution:

As Realm provide a method to know if a realm instance is in a transaction, we added a boolean variable to know if we need to commit the transaction after the block() depending if it was opened a transaction or not.

# Implemented roll-back transaction if a exception appear:

Problem:

The current implementation of the method [write:block] in the MJDRealmCacheService.m class doesn't implement a roll-back transaction if a database exception raise.

Solution:

Added a try catch, catching all the code inside the block(). if some exception raised, the code will cancel the transaction and generate a error if the transaction was the main one. If the code is in a nested transaction will raise the same exception to the parent one.